+++ lib/commands/report_command.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require_relative 'command'
+
+class ReportCommand < Command
+  def execute
+    $stdout.puts @robot.report
+  end
+end