+++ lib/commands/turn_left_command.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require_relative 'command'
+
+class TurnLeftCommand < Command
+  def execute
+    @robot.turn_left
+  end
+end