+++ lib/commands/turn_right_command.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require_relative 'command'
+
+class TurnRightCommand < Command
+  def execute
+    @robot.turn_right
+  end
+end