+++ lib/exceptions.rb
@@ -1,4 +1,7 @@
 # frozen_string_literal: true
 
 class MapSetupError < StandardError; end
+
 class InvalidMoveError < StandardError; end
+
+class MethodNotImplementedError < StandardError; end